feat: add new business rules

+++ src/core/delete-product.ts
@@ -1,7 +1,18 @@
 import { prisma } from "../infra/prisma";
+import { ResourceNotFoundError } from "./errors/resource-not-found-error";
 
 export class DeleteProductUseCase {
   async execute({ id }: { id: string }) {
+    const product = await prisma.product.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!product) {
+      throw new ResourceNotFoundError('Product not found.');
+    }
+
     await prisma.product.delete({
       where: {
         id,
